requirements.txt
Flask==3.1.1
# openai
python-dotenv==1.1.0
google-generativeai==0.8.5
smtplib
# email-validator
# sentence-transformers
Flask-SocketIO==5.5.1
# Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
eventlet==0.40.0
pdfplumber==0.11.7
python-docx==1.2.0

Flask-Login==0.6.3
# cohere==5.15.0
# APScheduler==3.11.0
Flask-APScheduler==1.13.1
